+++ main/dpa-verapdf/src/main/java/dk/statsbiblioteket/digital_pligtaflevering_aviser/verapdf/VeraPDFOutputValidation.java
@@ -1,6 +1,9 @@
 package dk.statsbiblioteket.digital_pligtaflevering_aviser.verapdf;
 
+import dk.kb.stream.StreamTuple;
+import io.vavr.control.Try;
 import org.w3c.dom.Document;
+import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 import org.xml.sax.SAXException;
 
@@ -14,8 +17,19 @@ import javax.xml.xpath.XPathFactory;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.function.Function;
+import java.util.stream.IntStream;
+
+import static java.util.stream.Collectors.counting;
+import static java.util.stream.Collectors.groupingBy;
+import static java.util.stream.Collectors.joining;
+import static java.util.stream.Collectors.mapping;
+import static java.util.stream.Collectors.toList;
 
 /**
  * Validator for checking the results from VeraPDF
@@ -23,6 +37,8 @@ import java.util.List;
  * https://sbprojects.statsbiblioteket.dk/display/DPAA/Projektdokumenter?preview=/15993252/31490419/pdfa-anbefaling.pdf
  */
 public class VeraPDFOutputValidation {
+
+    public static final String REMOVE_NEWLINES_REGEX = "[ \t]*\n[ \t]+";
 
     String expression;
 
@@ -33,6 +49,66 @@ public class VeraPDFOutputValidation {
         } else {
             expression = "//validationResult/assertions/assertion[@status='FAILED']/ruleId/@clause";
         }
+    }
+
+    public static String getReportFor(Document xmlDocument) {
+        XPath xPath = XPathFactory.newInstance().newXPath();
+        NodeList nodeList = null;
+        final String expression = "//status[text() = 'FAILED']/..";
+        try {
+            nodeList = (NodeList) xPath.compile(expression).evaluate(xmlDocument, XPathConstants.NODESET);
+        } catch (XPathExpressionException e) {
+            throw new RuntimeException(expression, e);
+        }
+
+        /*
+         <testAssertions>
+            <ordinal>3446</ordinal>
+            <ruleId>
+                <specification>ISO_19005_1</specification>
+                <clause>6.2.3</clause>
+                <testNumber>4</testNumber>
+            </ruleId>
+            <status>FAILED</status>
+            <message>If an uncalibrated colour space is used in a file then that file shall contain a PDF/A-1
+        OutputIntent, as defined in 6.2.2
+                </message>
+            <location>
+                <level>CosDocument</level>
+                <context>root/document[0]/pages[0](4 0 obj PDPage)/contentStream[0]/operators[4138]/fillCS[0]</context>
+            </location>
+        </testAssertions>
+        */
+
+        // https://stackoverflow.com/a/23361853/53897
+
+        Map<SeverenessLevel, List<StreamTuple<String, String>>> outcomes = IntStream.range(0, nodeList.getLength())
+                .mapToObj(nodeList::item)
+                .map((Node node) ->
+                        Try.of(() -> new StreamTuple<>(
+                                "" + xPath.compile("ruleId/clause/text()").evaluate(node, XPathConstants.STRING),
+                                ("" + xPath.compile(
+                                        "concat(ruleId/clause/text(), ': ', message/text(), ' [ ', location/context/text(), ' ]')"
+                                ).evaluate(node, XPathConstants.STRING)).replaceAll(REMOVE_NEWLINES_REGEX, " ")
+                        )).get())
+                .collect(groupingBy(st -> severenessFor(st.left()), mapping(st -> st, toList())));
+
+        Optional<SeverenessLevel> worstOutcome = outcomes.keySet().stream().max(Comparator.naturalOrder());
+
+        return outcomes.entrySet().stream()
+                .peek((Map.Entry<SeverenessLevel, List<StreamTuple<String, String>>> s) -> {
+                })
+                .sorted(Comparator.comparing(Map.Entry::getKey))
+                .map(entry -> entry.getKey() + "\n"
+                        + "------------------------------------------\n"
+                        + entry.getValue().stream()
+                        .collect(groupingBy(Function.identity(), counting())).entrySet().stream()
+                        .sorted(Comparator.comparing(Map.Entry::getKey))
+                        .map((Map.Entry<StreamTuple<String, String>, Long> entry2) ->
+                                entry2.getValue() + ": " + entry2.getKey().right())
+                        .collect(joining("\n"))
+                )
+                .collect(joining("\n\n"));
     }
 
     /**
@@ -77,20 +153,20 @@ public class VeraPDFOutputValidation {
 
         new HashSet<>(l).stream()
                 .map(item -> new ValidationResult(item, severenessFor(item)))
-                .filter(vr -> vr.getValidationEnum() != ValidationResult.ValidationResultEnum.ACCEPTABLE)
+                .filter(vr -> vr.getValidationEnum() != SeverenessLevel.ACCEPTABLE)
                 .forEach(vr -> rulesBroken.add(vr));
 
         return rulesBroken;
     }
 
-    public ValidationResult.ValidationResultEnum severenessFor(String sectionId) {
+    public static SeverenessLevel severenessFor(String sectionId) {
         switch (sectionId) {
             case "6.1.3":
             case "6.5.2":
             case "6.6.1":
             case "6.6.2":
             case "6.9":
-                return ValidationResult.ValidationResultEnum.INVALID;
+                return SeverenessLevel.INVALID;
             case "6.1.7":
             case "6.1.11":
             case "6.2.6":
@@ -100,7 +176,7 @@ public class VeraPDFOutputValidation {
             case "6.3.6":
             case "6.2.10":
             case "6.3.2":
-                return ValidationResult.ValidationResultEnum.MANUAL_INSPECTION;
+                return SeverenessLevel.MANUAL_INSPECTION;
             case "6.1.2":
             case "6.1.4":
             case "6.1.6":
@@ -120,12 +196,12 @@ public class VeraPDFOutputValidation {
             case "6.7.10":
             case "6.4":
             case "6.1.5":
-                return ValidationResult.ValidationResultEnum.ACCEPTABLE;
+                return SeverenessLevel.ACCEPTABLE;
             default:
                 if (sectionId.startsWith("6.2.3") || sectionId.startsWith("6.7") || sectionId.startsWith("6.8")) {
-                    return ValidationResult.ValidationResultEnum.ACCEPTABLE;
+                    return SeverenessLevel.ACCEPTABLE;
                 } else {
-                    return ValidationResult.ValidationResultEnum.UNKNOWN;
+                    return SeverenessLevel.UNKNOWN;
                 }
         }
     }